refactor: test order services
