+++ src/Service/BaseOrderManager.php
@@ -7,25 +7,21 @@ namespace Linio\SellerCenter\Service;
 use DateTimeInterface;
 use Linio\Component\Util\Json;
 use Linio\SellerCenter\Application\Parameters;
-use Linio\SellerCenter\Application\Security\Signature;
 use Linio\SellerCenter\Contract\OrderSortDirections;
 use Linio\SellerCenter\Contract\OrderSortFilters;
 use Linio\SellerCenter\Contract\OrderStatus;
 use Linio\SellerCenter\Exception\EmptyArgumentException;
 use Linio\SellerCenter\Exception\InvalidDomainException;
-use Linio\SellerCenter\Factory\RequestFactory;
 use Linio\SellerCenter\Factory\Xml\Order\FailureReasonsFactory;
 use Linio\SellerCenter\Factory\Xml\Order\OrderFactory;
 use Linio\SellerCenter\Factory\Xml\Order\OrderItemsFactory;
 use Linio\SellerCenter\Factory\Xml\Order\OrdersFactory;
 use Linio\SellerCenter\Factory\Xml\Order\OrdersItemsFactory;
 use Linio\SellerCenter\Factory\Xml\Order\TrackingCodeFactory;
-use Linio\SellerCenter\Formatter\LogMessageFormatter;
 use Linio\SellerCenter\Model\Order\FailureReason;
 use Linio\SellerCenter\Model\Order\Order;
 use Linio\SellerCenter\Model\Order\OrderItem;
 use Linio\SellerCenter\Model\Order\TrackingCode;
-use Linio\SellerCenter\Response\HandleResponse;
 use Linio\SellerCenter\Response\SuccessResponse;
 
 class BaseOrderManager extends BaseManager
@@ -35,59 +31,24 @@ class BaseOrderManager extends BaseManager
     public const DEFAULT_SORT_BY = 'created_at';
     public const DEFAULT_SORT_DIRECTION = 'ASC';
 
-    public function getOrder(int $orderId): Order
-    {
+    public function getOrder(
+        int $orderId
+        ): Order {
         $action = 'GetOrder';
 
-        $parameters = clone $this->parameters;
+        $parameters = $this->makeParametersForAction($action);
+
         $parameters->set([
-            'Action' => $action,
             'OrderId' => $orderId,
         ]);
-        $parameters->set([
-            'Signature' => Signature::generate($parameters, $this->configuration->getKey())->get(),
-        ]);
 
-        $requestHeaders = $this->generateRequestHeaders();
-        $requestId = $requestHeaders[self::REQUEST_ID_HEADER];
+        $requestId = $this->generateRequestId();
 
-        $request = RequestFactory::make(
-            'GET',
-            $this->configuration->getEndpoint(),
-            $requestHeaders
-        );
-
-        $this->logger->debug(
-            LogMessageFormatter::fromAction($requestId, $action, LogMessageFormatter::TYPE_REQUEST),
-            [
-                'url' => (string) $request->getUri(),
-                'method' => $request->getMethod(),
-                'body' => (string) $request->getBody(),
-                'parameters' => $parameters->all(),
-            ]
-        );
-
-        $response = $this->client->send($request, [
-            'query' => $parameters->all(),
-        ]);
-
-        $body = (string) $response->getBody();
-
-        $this->logger->debug(
-            LogMessageFormatter::fromAction($requestId, $action, LogMessageFormatter::TYPE_RESPONSE),
-            [
-                'body' => $body,
-            ]
-        );
-
-        $builtResponse = HandleResponse::parse($body);
-
-        $this->logger->debug(
-            LogMessageFormatter::fromAction($requestId, $action, LogMessageFormatter::TYPE_BUILT_RESPONSE),
-            [
-                'head' => $builtResponse->getHead()->asXML(),
-                'body' => $builtResponse->getBody()->asXML(),
-            ]
+        $builtResponse = $this->executeAction(
+            $action,
+            $parameters,
+            $requestId,
+            'GET'
         );
 
         $orderResponse = OrderFactory::make($builtResponse->getBody()->Orders->Order);
@@ -95,7 +56,7 @@ class BaseOrderManager extends BaseManager
         $this->logger->info(
             sprintf(
                 '%d::%s::APIResponse::SellerCenterSdk: the order was recovered',
-                $request->getHeaderLine('Request-ID'),
+                $requestId,
                 $action
             )
         );
@@ -106,11 +67,13 @@ class BaseOrderManager extends BaseManager
     /**
      * @return OrderItem[]
      */
-    public function getOrderItems(int $orderId): array
-    {
+    public function getOrderItems(
+        int $orderId
+        ): array {
         $action = 'GetOrderItems';
 
         $parameters = $this->makeParametersForAction($action);
+
         $parameters->set([
             'OrderId' => $orderId,
         ]);
@@ -145,8 +108,9 @@ class BaseOrderManager extends BaseManager
      *
      * @return Order[]
      */
-    public function getMultipleOrderItems(array $orderIdList): array
-    {
+    public function getMultipleOrderItems(
+        array $orderIdList
+        ): array {
         $action = 'GetMultipleOrderItems';
 
         $parameters = $this->makeParametersForAction($action);
@@ -154,6 +118,7 @@ class BaseOrderManager extends BaseManager
         if (empty($orderIdList)) {
             throw new EmptyArgumentException('OrderIdList');
         }
+
         $parameters->set([
             'OrderIdList' => Json::encode($orderIdList),
         ]);
@@ -186,9 +151,11 @@ class BaseOrderManager extends BaseManager
     /**
      * @return Order[]
      */
-    protected function getOrders(Parameters $parameters): array
-    {
+    protected function getOrders(
+        Parameters $parameters
+        ): array {
         $action = 'GetOrders';
+
         $parameters = $this->makeParametersForAction($action);
 
         $requestId = $this->generateRequestId();
@@ -425,6 +392,7 @@ class BaseOrderManager extends BaseManager
         string $shippingProvider
     ): TrackingCode {
         $action = 'GetTrackingCode';
+
         $parameters = $this->makeParametersForAction($action);
 
         $parameters->set([
@@ -433,6 +401,7 @@ class BaseOrderManager extends BaseManager
         ]);
 
         $requestId = $this->generateRequestId();
+
         $response = $this->executeAction(
             $action,
             $parameters,
@@ -452,13 +421,14 @@ class BaseOrderManager extends BaseManager
     }
 
     public function setStatusToCanceled(
-        int $orderItemId, 
-        string $reason, 
+        int $orderItemId,
+        string $reason,
         string $reasonDetail = null
         ): SuccessResponse {
         $action = 'SetStatusToCanceled';
 
         $parameters = $this->makeParametersForAction($action);
+
         $parameters->set([
             'OrderItemId' => $orderItemId,
             'Reason' => $reason,
